requirements.txt
streamlit
phidata
google-generativeai
duckduckgo-search
SpeechRecognition
PyAudio